Add Element.OuterHTML and Element.SetOuterHTML
The MDN marks these two as "experimental", but I cannot find any
actually authoritative information on this. innerHTML and outerHTML are
covered in the same spec (a "living standard"), which is available as a
candidate and a draft. innerHTML, however, is not marked as
experimental, so it's not clear at all why the two should be different.
outerHTML is supported by all modern browsers.