Change *Element parameters to Element (#49)
This was an unintentional oversight in PR #44. Element is an interface
rather than a struct, so no need to take a pointer.

This is a breaking API change that is warranted because it's a new API
and it's unlikely to have many/any users, since it's not very usable
in its current state.

Updates PR #44.
Fixes issue #48.