Make ImageData implement draw.Image interface
According to HTML standard, ImageData is defined to contain
non-alpha-premultiplied pixels [1][2]:

>	Pixel values must not be premultiplied by alpha.

As a result, we change ImageData to use NRGBA rather than RGBA.

ImageData, in nature, is similar to image.NRGBA.

We change its API slightly to be more like image.NRGBA. This allows it
to implement image.Image and draw.Image interfaces. That makes ImageData
more useful, since it can be passed directly to drawing algorithms that
work with image.Image or draw.Image. Skipping a temporary image.NRGBA
can improve performance significantly.

Also implement bounds checking in At and Set methods to be consistent
with image package. (It's still possible to access the underlying data
directly, if desired, via Data field.)

A downside of this change is the need to import image package, which
can increase binary size. The increase is small enough, and the
benefits of making ImageData more useful for idiomatic usage are deemed
to justify the cost.

References:

1.	https://html.spec.whatwg.org/multipage/canvas.html#pixel-manipulation
2.	https://stackoverflow.com/questions/23497925/how-can-i-stop-the-alpha-premultiplication-with-canvas-imagedata