Make ImageData implement draw.Image interface (#50)
Make ImageData implement draw.Image interface

ImageData, in nature, is similar to image.NRGBA.

We change its API slightly to be more like image.NRGBA. This allows it
to implement image.Image and draw.Image interfaces. That makes ImageData
more useful, since it can be passed directly to drawing algorithms that
work with image.Image or draw.Image.

Using ImageData directly and avoiding a temporary image.NRGBA can
improve performance significantly, since that way, there's no need
to copy all of the pixel values from one to the other.

The reason we make the change from using RGBA to using NRGBA is because
ImageData is defined to contain non-alpha-premultiplied pixels,
according to HTML standard [1] and additional sources [2]:

>	Pixel values must not be premultiplied by alpha.

Also implement bounds checking in At and Set methods to be consistent
with image package. (It's still possible to access the underlying data
directly, if desired, via Data field.)

A downside of this change is the need to import image package, which
can increase binary size. The increase is small enough, and the
benefits of making ImageData more useful for idiomatic usage are deemed
to justify the cost.

References:

1.	https://html.spec.whatwg.org/multipage/canvas.html#pixel-manipulation
2.	https://stackoverflow.com/questions/23497925/how-can-i-stop-the-alpha-premultiplication-with-canvas-imagedata