v2: change ClientRect to *Rect, document API stability
Update GetBoundingClientRect to return DOMRect instead of the legacy
ClientRect. A breaking API change was needed to make it return a pointer
rather than a value, so it was a good opportunity to also update the
type at the same time.

Document API stability; v2 is currently in alpha and more API changes
may be done before it exits alpha state.