v2: add dom package
This change implements the starting point of a new v2 major version
of the dom package.

It is now implemented on top of the syscall/js API, which is supported
by Go WebAssembly and GopherJS (as of version 1.12-2). The design goal
has been to stay as close as possible to v1 API to make the transition
easier to make. The only changes that were done were out of necessity:

- All struct fields with `js:"foo"` tags have been replaced with
  equivalent methods
- Underlying() returns js.Value instead of *js.Object
- AddEventListener() returns js.Func instead of func(*js.Object)
- GetBoundingClientRect() returns *Rect instead of ClientRect

Update GetBoundingClientRect to return DOMRect instead of the legacy
ClientRect. A breaking API change was needed to make it return a pointer
rather than a value, so it was a good opportunity to also update the
type at the same time.

There is a remaining TODO in window.RequestAnimationFrame method
to make it so that canceling the callback via CancelAnimationFrame
makes it so the js.Func wrapper's Release method is called. Arranging
this is difficult (it requires either having an internal map and sync
primitive for tracking js.Func by IDs, or changing the return value of
the RequestAnimationFrame method from int to something where the
js.Func and its Release method can be accessed easily.

The current implementation is functional but may end up using more
memory in some situations. It's not blocking for the initial alpha
release, so it remains a TODO for now.

Document API stability; v2 is currently in alpha and more API changes
may be done before it exits alpha state.

This implementation was based on work done by @yml in
Pull Request https://github.com/dominikh/go-js-dom/pull/59.

Fixes #57.

Co-authored-by: Yann Malet <yann.malet@gmail.com>
Co-authored-by: Dominik Honnef <dominik@honnef.co>
GitHub-Pull-Request: #69