rebase PointerEvent on top of MouseEvent
According to documentation on MDN¹, the PointerEvent interface
inherits properties from MouseEvent, which in turn inherits from
UIEvent, and that inherits from Event.

Embed MouseEvent in PointerEvent type in order to gain access to
additional properties in PointerEvent.

This is similar to what was done to WheelEvent in Pull Request #77.

¹ https://developer.mozilla.org/en-US/docs/Web/API/PointerEvent